Move PartitionsResource to V1.
