+++ kafka-rest/src/test/java/io/confluent/kafkarest/resources/v1/PartitionsResourceConsumeTest.java
@@ -1,4 +1,19 @@
-package io.confluent.kafkarest.unit;
+/*
+ * Copyright 2020 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.kafkarest.resources.v1;
 
 import static io.confluent.kafkarest.entities.EntityUtils.encodeBase64Binary;
 import static java.util.Collections.singletonList;
@@ -24,7 +39,6 @@ import io.confluent.kafkarest.entities.BinaryConsumerRecord;
 import io.confluent.kafkarest.entities.ConsumerRecord;
 import io.confluent.kafkarest.entities.EmbeddedFormat;
 import io.confluent.kafkarest.extension.InstantConverterProvider;
-import io.confluent.kafkarest.resources.PartitionsResource;
 import io.confluent.kafkarest.v2.KafkaConsumerManager;
 import io.confluent.rest.RestConfigException;
 import io.confluent.rest.validation.JacksonMessageBodyProvider;
